# backend/internal/interface/http/handler/admin.go
package handler

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// AdminHandler handles admin-related HTTP requests
type AdminHandler struct {
	// TODO: Add use case dependencies
}

// NewAdminHandler creates a new AdminHandler
func NewAdminHandler(useCase interface{}) *AdminHandler {
	return &AdminHandler{}
}

// GetDashboard retrieves dashboard data
// @Summary Get admin dashboard
// @Description Get dashboard statistics and metrics (Admin only)
// @Tags admin
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Router /admin/dashboard [get]
func (h *AdminHandler) GetDashboard(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Get dashboard - TODO"})
}

// ListUsers lists all users
// @Summary List all users
// @Description Get paginated list of all users (Admin only)
// @Tags admin
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param page query int false "Page number" default(1)
// @Param limit query int false "Items per page" default(20)
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Router /admin/users [get]
func (h *AdminHandler) ListUsers(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "List users - TODO"})
}

// DeleteUser deletes a user
// @Summary Delete user
// @Description Delete a user account (Admin only)
// @Tags admin
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "User ID"
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Router /admin/users/{id} [delete]
func (h *AdminHandler) DeleteUser(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Delete user - TODO"})
}

// ListComments lists all comments
// @Summary List all comments
// @Description Get paginated list of all comments (Admin only)
// @Tags admin
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param page query int false "Page number" default(1)
// @Param limit query int false "Items per page" default(20)
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Router /admin/comments [get]
func (h *AdminHandler) ListComments(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "List comments - TODO"})
}

// DeleteComment deletes a comment
// @Summary Delete comment
// @Description Delete a comment (Admin only)
// @Tags admin
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "Comment ID"
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Router /admin/comments/{id} [delete]
func (h *AdminHandler) DeleteComment(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Delete comment - TODO"})
}

// ListCategories lists all categories
// @Summary List all categories (Admin)
// @Description Get all categories for management (Admin only)
// @Tags admin
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Router /admin/categories [get]
func (h *AdminHandler) ListCategories(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "List categories - TODO"})
}

// CreateCategory creates a new category
// @Summary Create category
// @Description Create a new blog category (Admin only)
// @Tags admin
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param request body map[string]interface{} true "Category creation request"
// @Success 201 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Router /admin/categories [post]
func (h *AdminHandler) CreateCategory(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Create category - TODO"})
}

// UpdateCategory updates a category
// @Summary Update category
// @Description Update an existing category (Admin only)
// @Tags admin
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "Category ID"
// @Param request body map[string]interface{} true "Category update request"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Router /admin/categories/{id} [put]
func (h *AdminHandler) UpdateCategory(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Update category - TODO"})
}

// DeleteCategory deletes a category
// @Summary Delete category
// @Description Delete a category (Admin only)
// @Tags admin
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "Category ID"
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Router /admin/categories/{id} [delete]
func (h *AdminHandler) DeleteCategory(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Delete category - TODO"})
}

// ListTags lists all tags
// @Summary List all tags (Admin)
// @Description Get all tags for management (Admin only)
// @Tags admin
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Router /admin/tags [get]
func (h *AdminHandler) ListTags(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "List tags - TODO"})
}

// CreateTag creates a new tag
// @Summary Create tag
// @Description Create a new blog tag (Admin only)
// @Tags admin
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param request body map[string]interface{} true "Tag creation request"
// @Success 201 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Router /admin/tags [post]
func (h *AdminHandler) CreateTag(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Create tag - TODO"})
}

// UpdateTag updates a tag
// @Summary Update tag
// @Description Update an existing tag (Admin only)
// @Tags admin
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "Tag ID"
// @Param request body map[string]interface{} true "Tag update request"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Router /admin/tags/{id} [put]
func (h *AdminHandler) UpdateTag(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Update tag - TODO"})
}

// DeleteTag deletes a tag
// @Summary Delete tag
// @Description Delete a tag (Admin only)
// @Tags admin
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "Tag ID"
// @Success 200 {object} map[string]interface{}
// @Failure 401 {object} map[string]interface{}
// @Failure 403 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Router /admin/tags/{id} [delete]
func (h *AdminHandler) DeleteTag(c *gin.Context) {
	c.JSON(http.StatusNotImplemented, gin.H{"message": "Delete tag - TODO"})
}